Refactor package name from mediapipe to qali and update references
