+++ app/src/androidTest/java/com/qali/headup/ExampleInstrumentedTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.mediapipe.examples.facelandmarker
+package com.qali.headup
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -35,6 +35,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.google.mediapipe.examples.facelandmarker", appContext.packageName)
+        assertEquals("com.qali.headup", appContext.packageName)
     }
 }
