+++ app/src/test/java/com/qali/headup/ExampleUnitTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.mediapipe.examples.facelandmarker
+package com.qali.headup
 
 import org.junit.Test
